requirements.txt
fastapi[all]
uvicorn
websockets
uuid

psycopg2-binary==2.9.5
